+++ pkg/roachprod/vm/aws/aws.go
@@ -7,6 +7,7 @@
 package aws
 
 import (
+	"context"
 	_ "embed"
 	"encoding/json"
 	"fmt"
@@ -35,7 +36,11 @@ import (
 const ProviderName = "aws"
 
 // providerInstance is the instance to be registered into vm.Providers by Init.
-var providerInstance = &Provider{}
+var providerInstance = &Provider{
+	CLIProvider: vm.CLIProvider{
+		CLICommand: "aws",
+	},
+}
 
 //go:embed config.json
 var configJson []byte
@@ -62,6 +67,7 @@ func Init() error {
 
 	haveRequiredVersion := func() bool {
 		cmd := exec.Command("aws", "--version")
+		vm.MaybeLogCmd(config.Logger, cmd)
 		output, err := cmd.Output()
 		if err != nil {
 			return false
@@ -247,8 +253,8 @@ type ProviderOpts struct {
 
 // Provider implements the vm.Provider interface for AWS.
 type Provider struct {
-	// Profile to manage cluster in
-	Profile string
+	//// Profile to manage cluster in
+	//Profile string
 
 	// Path to json for aws configuration, defaults to predefined configuration
 	Config *awsConfig
@@ -259,6 +265,8 @@ type Provider struct {
 
 	// aws accounts to perform action in, used by gcCmd only as it clean ups multiple aws accounts
 	AccountIDs []string
+
+	vm.CLIProvider
 }
 
 func (p *Provider) SupportsSpotVMs() bool {
@@ -282,7 +290,7 @@ func (p *Provider) GetPreemptedSpotVMs(
 		}
 		args = append(args, vmList.ProviderIDs()...)
 		var describeInstancesResponse DescribeInstancesOutput
-		err = p.runJSONCommand(l, args, &describeInstancesResponse)
+		err = p.RunJSONCommand(l, args, &describeInstancesResponse)
 		if err != nil {
 			// if the describe-instances operation fails with the error InvalidInstanceID.NotFound,
 			// we assume that the instance has been preempted and describe-instances operation is attempted one hour after the instance termination
@@ -353,7 +361,7 @@ func (p *Provider) GetVMSpecs(
 		}
 		args = append(args, list.ProviderIDs()...)
 		var describeInstancesResponse DescribeInstancesOutput
-		err := p.runJSONCommand(l, args, &describeInstancesResponse)
+		err := p.RunJSONCommand(l, args, &describeInstancesResponse)
 		if err != nil {
 			return nil, errors.Wrapf(err, "error describing instances in region %s: ", region)
 		}
@@ -595,7 +603,7 @@ func (p *Provider) editLabels(
 		regionArgs = append(regionArgs, list.ProviderIDs()...)
 
 		g.Go(func() error {
-			_, err := p.runCommand(l, regionArgs)
+			_, err := p.RunCommand(context.Background(), l, regionArgs)
 			return err
 		})
 	}
@@ -694,6 +702,9 @@ func (p *Provider) Create(
 	if err := g.Wait(); err != nil {
 		return err
 	}
+	if config.DryRun {
+		return nil
+	}
 
 	return p.waitForIPs(l, names, regions, providerOpts)
 }
@@ -769,7 +780,7 @@ func (p *Provider) Delete(l *logger.Logger, vms vm.List) error {
 			if len(data.TerminatingInstances) > 0 {
 				_ = data.TerminatingInstances[0].InstanceID // silence unused warning
 			}
-			return p.runJSONCommand(l, args, &data)
+			return p.RunJSONCommand(l, args, &data)
 		})
 	}
 	return g.Wait()
@@ -790,8 +801,9 @@ func (p *Provider) Reset(l *logger.Logger, vms vm.List) error {
 		}
 		args = append(args, list.ProviderIDs()...)
 		g.Go(func() error {
-			_, e := p.runCommand(l, args)
-			return e
+			_, err := p.RunCommand(context.Background(), l, args)
+
+			return err
 		})
 	}
 	return g.Wait()
@@ -839,7 +851,8 @@ func (p *Provider) iamGetUser(l *logger.Logger) (string, error) {
 		}
 	}
 	args := []string{"iam", "get-user"}
-	err := p.runJSONCommand(l, args, &userInfo)
+	err := p.RunJSONCommand(l, args, &userInfo, true)
+
 	if err != nil {
 		return "", err
 	}
@@ -856,7 +869,8 @@ func (p *Provider) stsGetCallerIdentity(l *logger.Logger) (string, error) {
 		Arn string
 	}
 	args := []string{"sts", "get-caller-identity"}
-	err := p.runJSONCommand(l, args, &userInfo)
+	err := p.RunJSONCommand(l, args, &userInfo)
+
 	if err != nil {
 		return "", err
 	}
@@ -965,8 +979,8 @@ func (p *Provider) getVolumesForInstance(
 		"--region", region,
 		"--filters", "Name=attachment.instance-id,Values=" + instanceID,
 	}
+	err = p.RunJSONCommand(l, getVolumesArgs, &volumeOut)
 
-	err = p.runJSONCommand(l, getVolumesArgs, &volumeOut)
 	if err != nil {
 		return vols, err
 	}
@@ -1069,7 +1083,8 @@ func (p *Provider) listRegion(
 		"--region", region,
 	}
 	var describeInstancesResponse DescribeInstancesOutput
-	err := p.runJSONCommand(l, args, &describeInstancesResponse)
+	err := p.RunJSONCommand(l, args, &describeInstancesResponse)
+
 	if err != nil {
 		return nil, err
 	}
@@ -1163,6 +1178,78 @@ func (p *Provider) listRegion(
 	return ret, nil
 }
 
+func (p *Provider) validateMachineType(
+	l *logger.Logger,
+	machineType string,
+	region string,
+	ebsVolumes ebsVolumeList,
+	useLocalSSD bool,
+	useMultipleDisks bool,
+) error {
+	args := []string{
+		"ec2", "describe-instance-types",
+		"--instance-types", machineType,
+		"--region", region,
+	}
+	var data struct {
+		InstanceTypes []struct {
+			InstanceStorageSupported bool `json:"InstanceStorageSupported"`
+			InstanceStorageInfo      struct {
+				TotalSizeInGB float32 `json:"TotalSizeInGB"`
+				Disks         []struct {
+					SizeInGB float32 `json:"SizeInGB"`
+					Count    float32 `json:"Count"`
+					Type     string  `json:"Type"`
+				}
+			}
+			EbsInfo struct {
+				EbsOptimizedInfo struct {
+					BaselineBandwidthInMbps  float32 `json:"BaselineBandwidthInMbps"`
+					BaselineThroughputInMBps float32 `json:"BaselineThroughputInMBps"`
+					BaselineIOPS             float32 `json:"BaselineIOPS"`
+					MaximumBandwidthInMbps   float32 `json:"MaximumBandwidthInMbps"`
+					MaximumThroughputInMBps  float32 `json:"MaximumThroughputInMBps"`
+					MaximumIOPS              float32 `json:"MaximumIOPS"`
+				}
+			}
+		}
+	}
+	//_ = data.InstanceTypes // silence unused warning
+	//if len(data.InstanceTypes) > 0 {
+	//	_ = data.InstanceTypes[0].InstanceStorageSupported // silence unused warning
+	//}
+
+	if err := p.RunJSONCommand(l, args, &data, true); err != nil {
+		return err
+	}
+
+	if len(data.InstanceTypes) == 0 {
+		return nil
+	}
+	if !data.InstanceTypes[0].InstanceStorageSupported {
+		return nil
+	}
+	// TODO: use verbose when volumes are assigned
+	strs := make([]string, len(ebsVolumes))
+	for i, v := range ebsVolumes {
+		strs[i] = fmt.Sprintf("%+v", v)
+	}
+	fmt.Printf("EBS volumes: %s\n", strings.Join(strs, ", "))
+
+	if len(ebsVolumes) >= 2 && !useMultipleDisks {
+		fmt.Printf("using local NVMes with multiple EBS volumes without --aws-use-multiple-disks\n")
+		//return errors.Newf("instance type %s supports multiple disks, but --aws-use-multiple-disks was not specified", len(ebsVolumes)-1)
+	}
+	if !useLocalSSD {
+		fmt.Printf("using local NVMes without --aws-use-local-ssd\n")
+	}
+
+	pretty, _ := json.MarshalIndent(data.InstanceTypes[0].EbsInfo, "", "  ")
+	fmt.Println(string(pretty))
+
+	return nil
+}
+
 // runInstance is responsible for allocating a single ec2 vm.
 // Given that every AWS region may as well be a parallel dimension,
 // we need to do a bit of work to look up all of the various ids that
@@ -1243,9 +1330,12 @@ func (p *Provider) runInstance(
 	if err != nil {
 		return errors.Wrapf(err, "could not write AWS startup script to temp file")
 	}
-	defer func() {
-		_ = os.Remove(filename)
-	}()
+	// Keep startup script in dry-run or verbose mode.
+	if !config.DryRun && !config.Verbose {
+		defer func() {
+			_ = os.Remove(filename)
+		}()
+	}
 
 	withFlagOverride := func(cfg string, fl *string) string {
 		if *fl == "" {
@@ -1295,6 +1385,10 @@ func (p *Provider) runInstance(
 		args = append(args, "--iam-instance-profile", "Name="+p.IAMProfile)
 	}
 	ebsVolumes := assignEBSVolumes(&opts, providerOpts)
+
+	if err = p.validateMachineType(l, machineType, az.Region.Name, ebsVolumes, opts.SSDOpts.UseLocalSSD, providerOpts.UseMultipleDisks); err != nil {
+		return err
+	}
 	args, err = genDeviceMapping(ebsVolumes, args)
 	if err != nil {
 		return err
@@ -1305,7 +1399,7 @@ func (p *Provider) runInstance(
 		//todo(babusrithar): Add fallback to on-demand instances if spot instances are not available.
 	}
 	runInstancesOutput := RunInstancesOutput{}
-	return p.runJSONCommand(l, args, &runInstancesOutput)
+	return p.RunJSONCommand(l, args, &runInstancesOutput)
 }
 
 // runSpotInstance uses run-instances command to create a spot instance.
@@ -1319,7 +1413,7 @@ func runSpotInstance(l *logger.Logger, p *Provider, args []string, regionName st
 		fmt.Sprintf("MarketType=spot,SpotOptions={SpotInstanceType=one-time,"+
 			"InstanceInterruptionBehavior=terminate}"))
 	runInstancesOutput := RunInstancesOutput{}
-	err := p.runJSONCommand(l, spotArgs, &runInstancesOutput)
+	err := p.RunJSONCommand(l, spotArgs, &runInstancesOutput)
 	if err != nil {
 		return err
 	}
@@ -1372,7 +1466,7 @@ func cancelSpotRequest(
 		"--region", regionName,
 		"--spot-instance-request-ids", spotInstanceRequestId,
 	}
-	err := p.runJSONCommand(l, csrArgs, &CancelSpotInstanceRequestsOutput{})
+	err := p.RunJSONCommand(l, csrArgs, &CancelSpotInstanceRequestsOutput{})
 	if err != nil {
 		// This code path is not expected to be hit, but if it does, we should return the error, so that roachprod
 		// can destroy the cluster being created.
@@ -1391,7 +1485,7 @@ func describeSpotInstanceRequest(
 		"--spot-instance-request-ids", spotInstanceRequestId,
 	}
 	var describeSpotInstanceRequestsOutput DescribeSpotInstanceRequestsOutput
-	err := p.runJSONCommand(l, dsirArgs, &describeSpotInstanceRequestsOutput)
+	err := p.RunJSONCommand(l, dsirArgs, &describeSpotInstanceRequestsOutput)
 	if err != nil {
 		return DescribeSpotInstanceRequestsOutput{}, err
 	}
@@ -1433,7 +1527,7 @@ func getSpotInstanceRequestId(
 		"--instance-ids", instanceId,
 	}
 	var describeInstancesResponse DescribeInstancesOutput
-	err := p.runJSONCommand(l, diArgs, &describeInstancesResponse)
+	err := p.RunJSONCommand(l, diArgs, &describeInstancesResponse)
 	if err != nil {
 		return "", err
 	}
@@ -1529,7 +1623,8 @@ func (p *Provider) AttachVolume(l *logger.Logger, volume vm.Volume, vm *vm.VM) (
 	}
 
 	var commandResponse attachJsonResponse
-	err := p.runJSONCommand(l, args, &commandResponse)
+	err := p.RunJSONCommand(l, args, &commandResponse)
+
 	if err != nil {
 		return "", err
 	}
@@ -1546,7 +1641,8 @@ func (p *Provider) AttachVolume(l *logger.Logger, volume vm.Volume, vm *vm.VM) (
 		"--block-device-mappings",
 		"DeviceName=" + deviceName + ",Ebs={DeleteOnTermination=true,VolumeId=" + volume.ProviderResourceID + "}",
 	}
-	_, err = p.runCommand(l, args)
+	_, err = p.RunCommand(context.Background(), l, args)
+
 	if err != nil {
 		return "", err
 	}
@@ -1618,7 +1714,9 @@ func (p *Provider) CreateVolume(
 	}
 	args = append(args, "--size", strconv.Itoa(vco.Size))
 	var volumeDetails createVolume
-	err = p.runJSONCommand(l, args, &volumeDetails)
+
+	err = p.RunJSONCommand(l, args, &volumeDetails)
+
 	if err != nil {
 		return vol, err
 	}
@@ -1641,7 +1739,8 @@ func (p *Provider) CreateVolume(
 		"--query", "Volumes[*].State",
 	}
 	for waitForVolume.Next() {
-		err = p.runJSONCommand(l, args, &state)
+		err = p.RunJSONCommand(l, args, &state)
+
 		if len(state) > 0 && state[0] == "available" {
 			close(waitForVolumeCloser)
 		}
@@ -1704,9 +1803,8 @@ func (p *Provider) CreateVolumeSnapshot(
 		"--volume-id", volume.ProviderResourceID,
 		"--tag-specifications", "ResourceType=snapshot,Tags=[" + strings.Join(tags, ",") + "]",
 	}
-
 	var so snapshotOutput
-	if err := p.runJSONCommand(l, args, &so); err != nil {
+	if err := p.RunJSONCommand(l, args, &so); err != nil {
 		return vm.VolumeSnapshot{}, err
 	}
 	return vm.VolumeSnapshot{
